requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pynacl==1.5.0
pydantic==2.10.4
falkordb==1.0.9
httpx==0.28.1
